M11.3 fixups: CI clippy/fmt green + lights-buffer sized binding (review fix)
Two CI gates went red on the M11.3 increments and an adversarial review found one
real should-fix; all addressed here.

- workspace clippy (-D warnings, pedantic): env_import test cast (allow on the
  tiny fixed-size test buffer) + an elidable lifetime in lighting.rs.
- Tauri-app rustfmt: the inc.2b env_level0 match-chain was added after the last
  fmt; reformatted.
- review (LightBuf): the lights storage buffer is over-allocated (cap = next pow2)
  but was bound whole, so the shaders arrayLength(&lights) returned capacity, not
  the live count — fs_mesh looped trailing ZERO lights (a zero directional is
  normalize(vec3(0)) = NaN, sanitized by max() on this backend but a latent hazard
  + wasted per-fragment work). upload() now rebinds only the populated count*stride
  range, so arrayLength == the real light count.

Verified: cargo clippy --workspace + the Tauri-app manifest both clean under
-D warnings; cargo fmt --check clean; new tests green (lighting 2, env_import 2,
ibl 4); release exe rebuilt + the lights shader re-validates at runtime.

+++ assets/src/env_import.rs
@@ -43,6 +43,7 @@ mod tests {
     /// Encode a tiny HDR gradient (with an above-1.0 "sun" texel) to `.hdr`, decode it back, and check the
     /// dimensions + that the HDR highlight survived (Radiance RGBE is lossy, so values are approximate).
     #[test]
+    #[allow(clippy::cast_precision_loss)] // tiny fixed-size test buffer; the index casts are exact
     fn hdr_round_trips_dimensions_and_keeps_an_hdr_highlight() {
         let (w, h) = (4u32, 2u32);
         let mut buf = ImageBuffer::<Rgb<f32>, Vec<f32>>::new(w, h);
@@ -57,7 +58,7 @@ mod tests {
 
         let env = load_hdr_equirect(&bytes).expect("decode hdr");
         assert_eq!((env.width, env.height), (w, h));
-        assert_eq!(env.pixels.len() as u32, w * h);
+        assert_eq!(env.pixels.len(), (w * h) as usize);
         assert!(env.pixels[0][0] > 4.0, "the HDR sun survived (>1.0)");
     }
